Fix dependency cycle between util and config-utils
