+++ lib/analyze-action.js
@@ -54,7 +54,7 @@ async function sendStatusReport(startedAt, config, stats, error, trapCacheUpload
         ...statusReportBase,
         ...(config
             ? {
-                ml_powered_javascript_queries: util.getMlPoweredJsQueriesStatus(config),
+                ml_powered_javascript_queries: (0, config_utils_1.getMlPoweredJsQueriesStatus)(config),
             }
             : {}),
         ...(stats || {}),
