+++ lib/codeql.js
@@ -777,7 +777,7 @@ async function generateCodeScanningConfig(codeql, config, features, logger) {
     if (config.augmentationProperties.injectedMlQueries) {
         // We need to inject the ML queries into the original user input before
         // we pass this on to the CLI, to make sure these get run.
-        const packString = await util.getMlPoweredJsQueriesPack(codeql);
+        const packString = await (0, config_utils_1.getMlPoweredJsQueriesPack)(codeql);
         if (augmentedConfig.packs === undefined)
             augmentedConfig.packs = [];
         if (Array.isArray(augmentedConfig.packs)) {
