+++ lib/config-utils.test.js
@@ -35,6 +35,7 @@ const sinon = __importStar(require("sinon"));
 const api = __importStar(require("./api-client"));
 const codeql_1 = require("./codeql");
 const configUtils = __importStar(require("./config-utils"));
+const config_utils_1 = require("./config-utils");
 const feature_flags_1 = require("./feature-flags");
 const languages_1 = require("./languages");
 const logging_1 = require("./logging");
@@ -1456,4 +1457,64 @@ const mockRepositoryNwo = (0, repository_1.parseRepositoryNwo)("owner/repo");
         t.deepEqual(mockRequest.called, args.expectedApiCall);
     });
 });
+const ML_POWERED_JS_STATUS_TESTS = [
+    // If no packs are loaded, status is false.
+    [[], "false"],
+    // If another pack is loaded but not the ML-powered query pack, status is false.
+    [["some-other/pack"], "false"],
+    // If the ML-powered query pack is loaded with a specific version, status is that version.
+    [[`${config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.1.0`], "~0.1.0"],
+    // If the ML-powered query pack is loaded with a specific version and another pack is loaded, the
+    // status is the version of the ML-powered query pack.
+    [["some-other/pack", `${config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.1.0`], "~0.1.0"],
+    // If the ML-powered query pack is loaded without a version, the status is "latest".
+    [[config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME], "latest"],
+    // If the ML-powered query pack is loaded with two different versions, the status is "other".
+    [
+        [
+            `${config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.0.1`,
+            `${config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.0.2`,
+        ],
+        "other",
+    ],
+    // If the ML-powered query pack is loaded with no specific version, and another pack is loaded,
+    // the status is "latest".
+    [["some-other/pack", config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME], "latest"],
+];
+for (const [packs, expectedStatus] of ML_POWERED_JS_STATUS_TESTS) {
+    const packDescriptions = `[${packs
+        .map((pack) => JSON.stringify(pack))
+        .join(", ")}]`;
+    (0, ava_1.default)(`ML-powered JS queries status report is "${expectedStatus}" for packs = ${packDescriptions}`, (t) => {
+        return (0, util_1.withTmpDir)(async (tmpDir) => {
+            const config = {
+                languages: [],
+                queries: {},
+                paths: [],
+                pathsIgnore: [],
+                originalUserInput: {},
+                tempDir: tmpDir,
+                codeQLCmd: "",
+                gitHubVersion: {
+                    type: util_1.GitHubVariant.DOTCOM,
+                },
+                dbLocation: "",
+                packs: {
+                    javascript: packs,
+                },
+                debugMode: false,
+                debugArtifactName: util_1.DEFAULT_DEBUG_ARTIFACT_NAME,
+                debugDatabaseName: util_1.DEFAULT_DEBUG_DATABASE_NAME,
+                augmentationProperties: {
+                    injectedMlQueries: false,
+                    packsInputCombines: false,
+                    queriesInputCombines: false,
+                },
+                trapCaches: {},
+                trapCacheDownloadTime: 0,
+            };
+            t.is(configUtils.getMlPoweredJsQueriesStatus(config), expectedStatus);
+        });
+    });
+}
 //# sourceMappingURL=config-utils.test.js.map
