+++ lib/init-action.js
@@ -28,6 +28,7 @@ const core = __importStar(require("@actions/core"));
 const uuid_1 = require("uuid");
 const actions_util_1 = require("./actions-util");
 const api_client_1 = require("./api-client");
+const config_utils_1 = require("./config-utils");
 const environment_1 = require("./environment");
 const feature_flags_1 = require("./feature-flags");
 const init_1 = require("./init");
@@ -78,7 +79,7 @@ async function sendCompletedStatusReport(startedAt, config, toolsDownloadDuratio
             ...initStatusReport,
             disable_default_queries: disableDefaultQueries,
             languages,
-            ml_powered_javascript_queries: (0, util_1.getMlPoweredJsQueriesStatus)(config),
+            ml_powered_javascript_queries: (0, config_utils_1.getMlPoweredJsQueriesStatus)(config),
             paths,
             paths_ignore: pathsIgnore,
             queries: queries.join(","),
