+++ lib/util.test.js
@@ -141,69 +141,6 @@ const util = __importStar(require("./util"));
     t.is(util.apiVersionInRange("1.32.0", "1.33", "2.0"), util.DisallowedAPIVersionReason.ACTION_TOO_NEW);
     t.is(util.apiVersionInRange("2.1.0", "1.33", "2.0"), util.DisallowedAPIVersionReason.ACTION_TOO_OLD);
 });
-const ML_POWERED_JS_STATUS_TESTS = [
-    // If no packs are loaded, status is false.
-    [[], "false"],
-    // If another pack is loaded but not the ML-powered query pack, status is false.
-    [["some-other/pack"], "false"],
-    // If the ML-powered query pack is loaded with a specific version, status is that version.
-    [[`${util.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.1.0`], "~0.1.0"],
-    // If the ML-powered query pack is loaded with a specific version and another pack is loaded, the
-    // status is the version of the ML-powered query pack.
-    [
-        ["some-other/pack", `${util.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.1.0`],
-        "~0.1.0",
-    ],
-    // If the ML-powered query pack is loaded without a version, the status is "latest".
-    [[util.ML_POWERED_JS_QUERIES_PACK_NAME], "latest"],
-    // If the ML-powered query pack is loaded with two different versions, the status is "other".
-    [
-        [
-            `${util.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.0.1`,
-            `${util.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.0.2`,
-        ],
-        "other",
-    ],
-    // If the ML-powered query pack is loaded with no specific version, and another pack is loaded,
-    // the status is "latest".
-    [["some-other/pack", util.ML_POWERED_JS_QUERIES_PACK_NAME], "latest"],
-];
-for (const [packs, expectedStatus] of ML_POWERED_JS_STATUS_TESTS) {
-    const packDescriptions = `[${packs
-        .map((pack) => JSON.stringify(pack))
-        .join(", ")}]`;
-    (0, ava_1.default)(`ML-powered JS queries status report is "${expectedStatus}" for packs = ${packDescriptions}`, (t) => {
-        return util.withTmpDir(async (tmpDir) => {
-            const config = {
-                languages: [],
-                queries: {},
-                paths: [],
-                pathsIgnore: [],
-                originalUserInput: {},
-                tempDir: tmpDir,
-                codeQLCmd: "",
-                gitHubVersion: {
-                    type: util.GitHubVariant.DOTCOM,
-                },
-                dbLocation: "",
-                packs: {
-                    javascript: packs,
-                },
-                debugMode: false,
-                debugArtifactName: util.DEFAULT_DEBUG_ARTIFACT_NAME,
-                debugDatabaseName: util.DEFAULT_DEBUG_DATABASE_NAME,
-                augmentationProperties: {
-                    injectedMlQueries: false,
-                    packsInputCombines: false,
-                    queriesInputCombines: false,
-                },
-                trapCaches: {},
-                trapCacheDownloadTime: 0,
-            };
-            t.is(util.getMlPoweredJsQueriesStatus(config), expectedStatus);
-        });
-    });
-}
 (0, ava_1.default)("doesDirectoryExist", async (t) => {
     // Returns false if no file/dir of this name exists
     t.false(util.doesDirectoryExist("non-existent-file.txt"));
